Updated to latest LevelDB
 * The Windows port now uses Win32 native API for more reliability
 * Fixed various warnings
 * Removed useless "startthread" functions
 * Fixed corruption tests for Windows

+++ db/db_impl.cc
@@ -998,7 +998,7 @@ Status DBImpl::DoCompactionWork(CompactionState* compact) {
   CompactionStats stats;
   stats.micros = env_->NowMicros() - start_micros - imm_micros;
   for (int which = 0; which < 2; which++) {
-    for (int i = 0; i < compact->compaction->num_input_files(which); i++) {
+    for (size_t i = 0; i < compact->compaction->num_input_files(which); i++) {
       stats.bytes_read += compact->compaction->input(which, i)->file_size;
     }
   }
